WeaponBob Sprite Scaling
Tbh this feature can be completely ignored...that said, weapons scale up in size as bobrangeyY decreases in value. The intended effect is that weapon HUD sprites scale large in both x and y values as they move lower on the HUD, thus providing a sense of perspective, Feel free to check this out if you want Gen lol.